+++ extensions/gitlab/src/components/project.tsx
@@ -30,7 +30,7 @@ export function ProjectListItem(props: { project: Project }): JSX.Element {
     <List.Item
       id={project.id.toString()}
       title={project.name_with_namespace}
-      subtitle={"Stars " + project.star_count}
+      subtitle={project.star_count > 0 ? `⭐ ${project.star_count}` : ""}
       icon={localImageFilepath}
       actions={
         <ActionPanel>