+++ extensions/quick-access/CHANGELOG.md
@@ -1,5 +1,11 @@
 # Quick Access Changelog
 
+## [Update file preview] - 2022-05-19
+
+- Update file preview
+- Update "Show in Finder" action shortcuts
+- Add action "Open Command Preferences"
+
 ## [Update file preview] - 2022-04-30
 
 - Update file preview